Refactoring to a joke card and renaming module from common to shared
