--- src/app/common/common.module.ts
@@ -1,15 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { JokeCardComponent } from './joke-card/joke-card.component';
-
-
-
-@NgModule({
-  declarations: [
-    JokeCardComponent
-  ],
-  imports: [
-    CommonModule
-  ]
-})
-export class CommonModule { }
