--- src/app/common/joke-card/joke-card.component.html
@@ -1,1 +1,0 @@
-<p>joke-card works!</p>
