--- src/app/common/joke-card/joke-card.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-joke-card',
-  templateUrl: './joke-card.component.html',
-  styleUrls: ['./joke-card.component.scss']
-})
-export class JokeCardComponent {
-
-}
