+++ src/app/home/home-page/home-page.component.html
@@ -17,19 +17,8 @@
         <div class="box spinner-div"><mat-spinner></mat-spinner></div>
     </div>
     <div class="col-xs-12" *ngIf="jokesList.length >= MAX_JOKES_PER_PAGE">
-        <mat-card class="joke-card" *ngFor="let joke of jokesList.reverse(); let i = index">
-            <mat-card-content>
-                <div class="row">
-                    <div class="col-xs-10 middle-left">
-                        <span [ngClass]="{'fading': (i==0)}">{{joke.value}}</span>
-                    </div>
-                    <div class="col-xs middle-right">
-                        <button mat-icon-button aria-label="Add to my favorites jokes" (click)="addToFavorites(joke)">
-                            <mat-icon>favorite</mat-icon>
-                        </button>
-                    </div>
-                </div>
-            </mat-card-content>
-        </mat-card>
+        <ng-container *ngFor="let joke of jokesList.reverse(); let i = index">
+            <app-joke-card [index]="i" [joke]="joke"></app-joke-card>
+        </ng-container >
     </div>
 </div>
