+++ src/app/shared/joke-card/joke-card.component.html
@@ -1,0 +1,14 @@
+<mat-card class="joke-card">
+    <mat-card-content>
+        <div class="row">
+            <div class="col-xs-10 middle-left">
+                <span [ngClass]="{'fading': (index==0)}">{{joke?.value}}</span>
+            </div>
+            <div class="col-xs middle-right">
+                <button mat-icon-button aria-label="Add to my favorites jokes" (click)="addToFavorites(joke)">
+                    <mat-icon>favorite</mat-icon>
+                </button>
+            </div>
+        </div>
+    </mat-card-content>
+</mat-card>
