+++ src/app/shared/joke-card/joke-card.component.scss
@@ -1,0 +1,26 @@
+.joke-card {
+    margin: 10px;
+    padding: 20px;
+}
+.fading {
+    -webkit-animation-duration: 1s;
+    -webkit-animation-name: fading;
+    animation-duration: 1s;
+    animation-name: fading;
+}
+@-webkit-keyframes fading {
+    from {
+        color: green;
+    }
+    to {
+        color: white;
+    }
+}
+@keyframes fading {
+    from {
+        color: green;
+    }
+    to {
+        color: white;
+    }
+}
